+++ tombot/layer.py
@@ -3,7 +3,6 @@ import os
 import sys
 import logging
 import time
-import re
 import sqlite3
 import datetime
 import threading
@@ -46,10 +45,6 @@ class TomBotLayer(YowInterfaceLayer):
         except KeyError:
             logging.critical('Database could not be loaded!')
 
-        # Mentioning setup
-        mention_regex = r'(?<!\w)@\s?(\w+)[ .:,]?'
-        self.mention_pattern = re.compile(mention_regex)
-
         # Group list holder
         self.known_groups = []
 
@@ -69,7 +64,6 @@ class TomBotLayer(YowInterfaceLayer):
             'HELP'      : self.help,
             'REMINDME'  : self.addreminder,
             'REMIND'    : self.addreminder,
-            'BOTHER'    : self.anonsend,
             }
         plugins.load_plugins()
         self.functions.update(plugins.COMMANDS)
@@ -122,48 +116,8 @@ class TomBotLayer(YowInterfaceLayer):
         time.sleep(0.2)
         self.react(message)
 
-        # Handle mentions:
-        mentioned_sent = []
-        # Used to not send multiple messages to the same person if multiple
-        # nicks resolve to the same jid
-        for nick in self.mention_pattern.findall(message.getBody()):
-            logging.debug('Nick detected: %s', nick)
-
-            # Who sent the message?
-            senderjid = determine_sender(message)
-            logging.debug('Resolving sender %s', senderjid)
-            try:
-                sendername = self.jid_to_nick(senderjid)
-                logging.debug('Sendernick %s', sendername)
-            except KeyError:
-                sendername = senderjid
-                logging.debug('Could not find jid %s', senderjid)
-
-            # Who was mentioned?
-            try:
-                targetjid = self.nick_to_jid(nick)
-            except KeyError as e:
-                # Some nick that is unknown, pass
-                logging.debug('Could not resolve nick %s', nick)
-                logging.debug('Exception %s', e)
-                continue
-
-            if targetjid not in mentioned_sent:
-                # Check timeout
-                t_info = self.get_jid_timeout(targetjid)
-                currenttime = (datetime.datetime.now() - datetime.datetime(
-                    1970, 1, 1)).total_seconds()
-                if currenttime < (t_info[1] + t_info[0]) and message.participant:
-                    # Do not send DM if recipient has not timed out yet
-                    continue
-
-                # Send mention notification: [author]: [body]
-                entity = TextMessageProtocolEntity('{}: {}'.format(
-                    sendername, message.getBody()), to=targetjid)
-                self.toLower(entity)
-                mentioned_sent.append(targetjid)
-        # Updating user's last seen is after mentions so mention timeouts can be tested solo
-        self.update_lastseen(message)
+        for handler in plugins.MESSAGE_HANDLERS:
+            handler(self, message)
 
     @ProtocolEntityCallback('receipt')
     def onReceipt(self, entity):
@@ -211,29 +165,6 @@ class TomBotLayer(YowInterfaceLayer):
                 response.encode('utf-8'), to=message.getFrom())
             self.toLower(reply_message)
 
-    def anonsend(self, message):
-        ''' Send a mention under the group name and not the author's name '''
-        if not message.participant:
-            return
-
-        try:
-            groupname = self.jid_to_nick(message.getFrom())
-        except KeyError:
-            return 'This group is not enrolled in the BrotherBother program, sorry'
-
-        text = extract_query(message, 2)
-        body = '{}: {}'.format(groupname, text)
-
-        # Who was mentioned?
-        nick = message.getBody().split()[2]
-        try:
-            recipient = self.nick_to_jid(nick)
-        except KeyError:
-            return 'Unknown recipient!'
-
-        entity = TextMessageProtocolEntity(body, to=recipient)
-        self.toLower(entity)
-
     def help(self, message):
         ''' TODO: give an overview of available commands. '''
         # pylint: disable=unused-argument
@@ -251,16 +182,6 @@ class TomBotLayer(YowInterfaceLayer):
         if restart:
             sys.exit(3)
         sys.exit(0)
-
-    # NewNicks
-    def update_lastseen(self, message):
-        ''' Update a user's last seen time in the database. '''
-        author = determine_sender(message)
-        currenttime = (datetime.datetime.now() - datetime.datetime(1970, 1, 1)).total_seconds()
-        logging.debug('Updating %s\'s last seen.', author)
-        self.cursor.execute('UPDATE users SET lastactive = ?, message = ? WHERE jid = ?',
-                            (currenttime, message.getBody().decode('utf-8'), author))
-        self.conn.commit()
 
     def add_other_nick(self, message):
         ''' TODO: Link a nickname to another user (admin-only) '''
@@ -304,56 +225,6 @@ class TomBotLayer(YowInterfaceLayer):
             [body, determine_sender(message)],
             run_date=deadline)
         return
-
-    def nick_to_jid(self, name):
-        '''
-        Maps a (nick)name to a jid using either users or nicks.
-
-        Raises KeyError if the name is unknown.
-        '''
-        # Search authornames first
-        queries = [
-            'SELECT jid FROM users WHERE primary_nick LIKE ?',
-            'SELECT jid FROM nicks WHERE name LIKE ?',
-            ]
-        for query in queries:
-            self.cursor.execute(query, (name,))
-            result = self.cursor.fetchone()
-            if result:
-                return result[0]
-
-        raise KeyError('Unknown nick {}!'.format(name))
-
-    def jid_to_nick(self, jid):
-        '''
-        Map a jid to the user's primary_nick.
-
-        Raises KeyError if user not known.
-        '''
-        query = 'SELECT primary_nick FROM users WHERE jid = ?'
-        self.cursor.execute(query, (jid,))
-        result = self.cursor.fetchone()
-        if result:
-            return result[0]
-
-        raise KeyError('Unknown jid {}'.format(jid))
-
-    def get_jid_timeout(self, jid):
-        '''
-        Retrieve a user's lastactive and timeout.
-
-        Returns a (timeout, lastactive) tuple.
-        Raises KeyError if jid not known.
-        '''
-        self.cursor.execute(
-            'SELECT timeout, lastactive FROM users WHERE jid = ?',
-            (jid,))
-        result = self.cursor.fetchone()
-
-        if result:
-            return result
-
-        raise KeyError('Unknown jid {}'.format(jid))
 
     # Helper functions
     def set_online(self, *_):
