+++ tombot/plugins/__init__.py
@@ -3,7 +3,8 @@ Provides the plugin infrastructure and some helper functions for plugins.
 '''
 import os.path
 import importlib
-from .registry import COMMANDS, get_easy_logger, STARTUP_FUNCTIONS, SHUTDOWN_FUNCTIONS
+from .registry import COMMANDS, STARTUP_FUNCTIONS, SHUTDOWN_FUNCTIONS, MESSAGE_HANDLERS
+from .registry import get_easy_logger
 
 
 LOGGER = get_easy_logger('moduleloader')
